Replace 'MUST' with 'SHOULD' for branch rule
Problem: The 'MUST' language carries lots of weight, and should only be
used for fundamentals like 'governance patches MUST NOT be merged
without approval'. Having multiple branches in a repository isn't an
ideal, and we should try to remove them when we can, but it isn't the
end of the world if a maintainer uses it for a pull request.

Solution: Reduce 'MUST' to 'SHOULD' for the rule that says that
repositories shouldn't have any other branches. After this change, the
only 'MUST' rules are fundamentals that are absolutely necessary.