Remove contributor PR update requirement
Problem: I don't think that we can make a requirement (especially a
'MUST' requirement) that contributors follow maintainer suggestions in
their patch.

Solution: We could add a "or close their pull request" suffix, but since
maintainers are tasked with closing out-of-date pull requests I don't
think that we need to put that responsibility on contributors. My
preferred solution would be to remove this requirement from
contributors. This commit also rearranges the lines so that the
harassment policy isn't in the middle of the PR policy.