Default to Prettier and ESlint
Our modules don't conform to any style guide for code formatting or
code quality, which means we all pull toward our personal preference
rather than some 'good enough' middle ground that we can all agree on.

Common style guides make cooperation easier and create a high-pass
filter for disagreements and churn. Automatic code formatting and
code-quality rules are helpful to set expectations that are accessible
and predictable.

Writing a style guide takes lots of work, and there's a large risk that
any custom style guide will be unfamiliar to the vast majority of people
who want to contribute to our projects. My preference is to use style
guides which are popular (familiar) and uncontroversial (boring), which
should make the style guide resistant to unnecessary future churn.

This commit adds a rule to default to Prettier and ESLint, which gives
us code formatting rules for most languages and code-quality rules for
languages that are compatible with ESLint (JavaScript, JSX, TypeScript,
and others).

Fixes: https://github.com/ssb-js/ssb-js/issues/25